+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/changelog.py
@@ -4,10 +4,10 @@
 import os
 from collections import namedtuple
 from datetime import datetime
+from io import StringIO
 
 import click
 from semver import parse_version_info
-from six import StringIO
 
 from ....utils import stream_file_lines, write_file
 from ...constants import CHANGELOG_TYPE_NONE, get_root
@@ -34,15 +34,15 @@ def changelog(ctx, check, version, old_version, initial, quiet, dry_run):
     This method is supposed to be used by other tasks and not directly.
     """
     if check and check not in get_valid_checks():
-        abort('Check `{}` is not an Agent-based Integration'.format(check))
+        abort(f'Check `{check}` is not an Agent-based Integration')
 
     # sanity check on the version provided
     cur_version = old_version or get_version_string(check)
     if parse_version_info(version.lstrip('v')) <= parse_version_info(cur_version.lstrip('v')):
-        abort('Current version is {}, cannot bump to {}'.format(cur_version, version))
+        abort(f'Current version is {cur_version}, cannot bump to {version}')
 
     if not quiet:
-        echo_info('Current version of check {}: {}, bumping to: {}'.format(check, cur_version, version))
+        echo_info(f'Current version of check {check}: {cur_version}, bumping to: {version}')
 
     # get the name of the current release tag
     target_tag = get_release_tag_string(check, cur_version)
@@ -53,7 +53,7 @@ def changelog(ctx, check, version, old_version, initial, quiet, dry_run):
     # for each PR get the title, we'll use it to populate the changelog
     pr_numbers = parse_pr_numbers(diff_lines)
     if not quiet:
-        echo_info('Found {} PRs merged since tag: {}'.format(len(pr_numbers), target_tag))
+        echo_info(f'Found {len(pr_numbers)} PRs merged since tag: {target_tag}')
 
     if initial:
         # Only use the first one
@@ -65,26 +65,26 @@ def changelog(ctx, check, version, old_version, initial, quiet, dry_run):
         try:
             payload = get_pr(pr_num, user_config)
         except Exception as e:
-            echo_failure('Unable to fetch info for PR #{}: {}'.format(pr_num, e))
+            echo_failure(f'Unable to fetch info for PR #{pr_num}: {e}')
             continue
 
         changelog_labels = get_changelog_types(payload)
 
         if not changelog_labels:
-            abort('No valid changelog labels found attached to PR #{}, please add one!'.format(pr_num))
+            abort(f'No valid changelog labels found attached to PR #{pr_num}, please add one!')
         elif len(changelog_labels) > 1:
-            abort('Multiple changelog labels found attached to PR #{}, please only use one!'.format(pr_num))
+            abort(f'Multiple changelog labels found attached to PR #{pr_num}, please only use one!')
 
         changelog_type = changelog_labels[0]
         if changelog_type == CHANGELOG_TYPE_NONE:
             if not quiet:
                 # No changelog entry for this PR
-                echo_info('Skipping PR #{} from changelog due to label'.format(pr_num))
+                echo_info(f'Skipping PR #{pr_num} from changelog due to label')
             continue
 
         author = payload.get('user', {}).get('login')
         author_url = payload.get('user', {}).get('html_url')
-        title = '[{}] {}'.format(changelog_type, payload.get('title'))
+        title = f"[{changelog_type}] {payload.get('title')}"
 
         entry = ChangelogEntry(pr_num, title, payload.get('html_url'), author, author_url, from_contributor(payload))
 
@@ -94,7 +94,7 @@ def changelog(ctx, check, version, old_version, initial, quiet, dry_run):
     new_entry = StringIO()
 
     # the header contains version and date
-    header = '## {} / {}\n'.format(version, datetime.utcnow().strftime('%Y-%m-%d'))
+    header = f"## {version} / {datetime.utcnow().strftime('%Y-%m-%d')}\n"
     new_entry.write(header)
 
     # one bullet point for each PR
@@ -102,8 +102,8 @@ def changelog(ctx, check, version, old_version, initial, quiet, dry_run):
     for entry in entries:
         thanks_note = ''
         if entry.from_contributor:
-            thanks_note = ' Thanks [{}]({}).'.format(entry.author, entry.author_url)
-        new_entry.write('* {}. See [#{}]({}).{}\n'.format(entry.title, entry.number, entry.url, thanks_note))
+            thanks_note = f' Thanks [{entry.author}]({entry.author_url}).'
+        new_entry.write(f'* {entry.title}. See [#{entry.number}]({entry.url}).{thanks_note}\n')
     new_entry.write('\n')
 
     # read the old contents