+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/test.py
@@ -16,9 +16,9 @@ from .console import CONTEXT_SETTINGS, abort, echo_info, echo_success, echo_wait
 
 def display_envs(check_envs):
     for check, envs in check_envs:
-        echo_success('`{}`:'.format(check))
+        echo_success(f'`{check}`:')
         for e in envs:
-            echo_info('    {}'.format(e))
+            echo_info(f'    {e}')
 
 
 @click.command(context_settings=CONTEXT_SETTINGS, short_help='Run tests')
@@ -106,16 +106,16 @@ def test(
     }
 
     if passenv:
-        test_env_vars['TOX_TESTENV_PASSENV'] += ' {}'.format(passenv)
+        test_env_vars['TOX_TESTENV_PASSENV'] += f' {passenv}'
 
-    test_env_vars['TOX_TESTENV_PASSENV'] += ' {}'.format(' '.join(get_ci_env_vars()))
+    test_env_vars['TOX_TESTENV_PASSENV'] += f" {' '.join(get_ci_env_vars())}"
 
     if color is not None:
         test_env_vars['PY_COLORS'] = '1' if color else '0'
 
     if e2e:
         test_env_vars[E2E_PARENT_PYTHON] = sys.executable
-        test_env_vars['TOX_TESTENV_PASSENV'] += ' {}'.format(E2E_PARENT_PYTHON)
+        test_env_vars['TOX_TESTENV_PASSENV'] += f' {E2E_PARENT_PYTHON}'
 
     check_envs = get_tox_envs(checks, style=style, format_style=format_style, benchmark=bench, changed_only=changed)
     tests_ran = False
@@ -153,7 +153,7 @@ def test(
         test_env_vars['PYTEST_ADDOPTS'] = pytest_options
 
         if verbose:
-            echo_info('pytest options: `{}`'.format(test_env_vars['PYTEST_ADDOPTS']))
+            echo_info(f"pytest options: `{test_env_vars['PYTEST_ADDOPTS']}`")
 
         with chdir(os.path.join(root, check), env_vars=test_env_vars):
             if format_style:
@@ -167,7 +167,7 @@ def test(
             else:
                 test_type_display = 'tests'
 
-            wait_text = '{}Running {} for `{}`'.format(output_separator, test_type_display, check)
+            wait_text = f'{output_separator}Running {test_type_display} for `{check}`'
             echo_waiting(wait_text)
             echo_waiting('-' * len(wait_text))